Exit early in a few places
The recent economic downturn has triggered a contraction in the global
supply of tab characters, causing prices to soar. God knows research
budgets aren't going to get increased any time soon, so cutting back
on expensive tab characters is necessary. One way to do this is to
exit early out of functions when possible.

For example, consider the following function:

```python
def function(test_condition):
	if test_condition:
		statement_1()
		statement_2()
		# ...
		statement_n_minus_1()
		return n
	else:
		return None
```

The `else` block is shorter than the `if` block, so the two can be
inverted by negating the test condition:

```python
def function(test_condition):
	if not test_condition:
		return None
	else:
		statement_1()
		statement_2()
		# ...
		statement_n_minus_1()
		return n
```

Next, the `if` block returns, so if it triggers, the rest of the
function cannot be reached. Therefore the `else` is not needed at all,
and the rest of the `else` block can be un-indented by one level:

```python
def function(test_condition):
	if not test_condition:
		return None

	statement_1()
	statement_2()
	# ...
	statement_n_minus_1()
	return n
```

This change reduces tab usage by n + 1, which at current market prices
is a tidy sum.